// src/types.ts
export type UserRole = 
  | 'developer'
  | 'super_admin'
  | 'admin'
  | 'operations'
  | 'finance'
  | 'merchant'
  | 'driver'
  | 'support'
  | 'customer';